requirements.txt
pynput>=1.7.6
sounddevice>=0.4.6
numpy>=1.24.0
openai-whisper>=20231117
PyYAML>=6.0.1
python-xlib>=0.33
pyautogui>=0.9.54
scipy>=1.11.0 